requirements.txt
# Core Python Libraries
numpy==2.3.0
pandas==2.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2

# FastAPI Framework
fastapi==0.115.12
uvicorn==0.34.3
starlette==0.46.2

# Supabase & Web
supabase
requests==2.32.4
python-multipart
anyio==4.9.0
h11==0.16.0
sniffio==1.3.1
httpx==0.27.2

# Type Checking & Validation
pydantic==2.11.6
pydantic_core==2.33.2
typing-inspection==0.4.1
typing_extensions==4.14.0
annotated-types==0.7.0

# Encoding & Networking
charset-normalizer==3.4.2
idna==3.10
urllib3==2.4.0
certifi==2025.4.26

# CLI / UX
click==8.2.1
colorama==0.4.6
six==1.17.0

python-dotenv
pdfplumber==0.11.4

# ML inference dependencies
joblib==1.4.2
scikit-learn==1.5.2
xgboost==2.1.1
lightgbm==4.5.0

# iRacing SDK (선택사항)
# pyirsdk  # 실제로 작동하는지 확인 필요
pywin32==306; sys_platform == "win32"  # Windows 공유 메모리 접근용